+++ src/app/store.ts
@@ -1,0 +1,11 @@
+import { configureStore } from '@reduxjs/toolkit';
+import authSlice from '../features/auth/authSlice';
+
+export const store = configureStore({
+  reducer: {
+    auth: authSlice,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
